pyproject.toml
[project]
name = "my-django-project-001"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2",
    "djangorestframework>=3.16.0",
    "dotenv>=0.9.9",
    "pandas>=2.2.3",
    "pymysql>=1.1.1",
    "sqlalchemy",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.4",
]
